Refactor reimbursement calculation logic and update domain models
- Change `Invoice.amount` from `Int` to `Long` and update related models for consistency.
- Change `Reimbursement.totalAmount` and `to` values from `Int` to `Float`.
- Extend in-memory tests with new sample invoices and additional test cases for reimbursement scenarios.
- Refactor `cleanupStorage` to delete all invoices and log completion.
- Improve reimbursement proposal logic in `InvoiceServiceLive` for accurate distribution calculations.

+++ domain/src/scala/domain/models/invoice/InvoiceCreate.scala
@@ -6,7 +6,7 @@ import java.time.LocalDate
 import scala.jdk.CollectionConverters.*
 
 case class InvoiceCreate(
-  amount: Int,
+  amount: Long,
   mileage: Option[Int],
   date: LocalDate,
   name: String,
